Split settings UI into partial files for more convenient development and maintainability
